all: use ROLLUP_WATCH instead of NODE_ENV
We want to enable the livereload plugin when running in watch mode.
We don't want to enable the livereload plugin when building, because
that causes the build to hang forever.

NODE_ENV is the wrong tool for the job here: it can be set to
"development" even when running `npm run build`, for instance if
the user has set it globally (on a dev machine). In general, NODE_ENV
is just supposed to indicate whether a Node.js app is running in
production - nothing else.

Instead, use the ROLLUP_WATCH env var: it's set by rollup specifically
when started in watch mode.

Signed-off-by: Simon Ser <contact@emersion.fr>