ui-manchette: fix manchette toolbar not fixed bottom
The PR https://github.com/OpenRailAssociation/osrd-ui/pull/728 changed the height of the manchette from a min height to a fixed one.
It introduced a bug where, when having a long list of waypoint, when scrolling bottom, the manchette toolbar wasn't stick to the bottom anymore.
Fix this issue by giving back a min height to the manchette while still keeping its dynamic value introduced in PR #728.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>