ui-manchette: fix waypoint menu not closing
As some of the manchette action buttons can be disabled when the waypoint menu is displayed, the click event when clicking on them wasn't fired and the menu would'nt close.
Update the style of the buttons to be able to click on the span inside the buttons which isn't disabled and update the story to listen to these click events in order to properly close the menu.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>