build: use rollup --failAfterWarnings
We've got to a point where we've eliminated almost all build errors.
We previously did not notice these because they only show up as
warnings in CI, without making CI fail.

Use rollup --failAfterWarnings so that CI fails when a build error
is emitted.

Leave ui-spacetimechart and ui-trackoccupancydiagram alone, because
these still have a failure due to a circular dependency:

    (!) [plugin typescript] ../ui-spacetimechart/src/lib/types.ts (6:8): @rollup/plugin-typescript TS2307: Cannot find module '@osrd-project/ui-trackoccupancydiagram/dist/components/types' or its corresponding type declarations.
    /home/runner/work/osrd-ui/osrd-ui/ui-spacetimechart/src/lib/types.ts:6:8

    6 } from '@osrd-project/ui-trackoccupancydiagram/dist/components/types';
             ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Simon Ser <contact@emersion.fr>