ui-manchette-with-spacetimechart: add ManchetteWithSpaceTimeChart
This is a component designed for simple use-cases. It's easier to
use than the hook. Users can still decide to use the hook instead
of this component, especially when they need more customizability.

Other components can be passed in for the header and SpaceTimeChart,
and additional properties can be passed to Manchette/SpaceTimeChart.

This is a breaking API change: the hook is no longer the default
export for ui-manchette-with-spacetimechart. Users need to replace:

    import useManchetteWithSpaceTimeChart from 'ui-manchette-with-spacetimechart';

with:

    import { useManchetteWithSpaceTimeChart } from 'ui-manchette-with-spacetimechart';

Signed-off-by: Simon Ser <contact@emersion.fr>